backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
requests==2.31.0
python-dotenv==1.0.1
pyodbc==5.1.0
pandas==2.1.4
selenium==4.15.2
webdriver-manager==4.0.1
cryptography==42.0.5
schedule==1.2.0
openpyxl==3.1.2
Pillow==10.2.0

